Rename files and directories so its in kebab case
