+++ packages/domain-search/src/components/domains-full-cart-items/style.scss
