+++ packages/domain-search/src/components/domains-full-cart/style.scss
