+++ packages/domain-search/src/hooks/use-click-outside.ts
